refactor: move plugin services to App\Services\Plugins
