+++ app/Services/Themes/PluginInformationService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\Themes;
 
 use App\Models\WpOrg\Plugin;
 use Illuminate\Database\Eloquent\Builder;
